pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zoom-ai"
version = "0.1.0"
description = "Virtual Avatar Meeting Assistant for Zoom"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["zoom", "virtual-avatar", "ai-assistant", "meeting-bot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
]

dependencies = [
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "edge-tts>=6.1.0",
    "playwright>=1.40.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.35.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "loguru>=0.7.0",
    "httpx>=0.25.0",
    "ffmpeg-python>=0.2.0",
    "av>=11.0.0",
    "aiohttp>=3.9.0",
    "pyvirtualcam>=0.6.0; platform_system=='Darwin'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
sadtalker = [
    "scipy>=1.10.0",
    "yacs>=0.1.8",
    "pyyaml>=6.0",
]
musetalk = [
    "diffusers>=0.24.0",
    "accelerate>=0.25.0",
]

[project.scripts]
zoom-ai = "zoom_ai.cli:main"

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[tool.setuptools.packages.find]
include = ["zoom_ai*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
